Change default from MPI_[I]Rsends to MPI_[I]sends
On some systems MPI_Irsends are buggy and can cause a hang (we
found this on titan with pio1). So switching the default from
using MPI_Irsends to MPI_Isends.

The user can use "_USE_MPI_RSEND" (compile time flag) to choose
MPI_[I]Rsends instead of the default

Fixes #946