Reading txt in header as strings
Reading the text information in the header of the io decomposition
dump file as strings (instead of matching the exact string written
out in PIOc_writemap). This allows a more flexible formatting for
the text (version string etc) in the header.

This change allows header files written out from Fortran to be
read (correctly) by the C lib.