Generalize the list read util function
Generalize the function that parses a list into an integer array
to accept character arrays. Now the same function can be used
to either parse a list of ints or a list of strings.

Note that currently we only accept either an integer array or
a character array (if both are provided we only parse inputs to
the integer array)