Reset NetCDF classic flags for NetCDF4 file types
When creating NetCDF4 files the NetCDF library does not allow
NC_64BIT_OFFSET and NC_64BIT_DATA flags. However these flags are
required to create CDF2 and CDF5 files respectively for NetCDF
classic files. Since these flags are passed (via PIO flags) it
is convenient for the user to be able to use the same flags for
all iotypes.