Alloc iobuf even when we write all fillvals
When allocating memory for iobuf don't look at the llen, that
can be 0 when we write all fillvalues from the local process.

Without this fix a test that writes out only fillvalues from any
MPI process would fail during write with NetCDF iotype (Since
iobuf, that is used for gathering data, is not allocated
because llen = 0 on root process and maxiobuflen != 0).