Use a more reasonable maxblock value for the box rearranger
In CalcStartandCount() for the box rearranger, maxbytes only
depends on DEFAULT_BLOCKSIZE. It should be calculated based
on the global size and number of ioprocs that are needed.

Without this fix, CalcStartandCount() might assign the whole
data to the first IO process, which does not give us an even
distribution.

Also updated some comments.