Adding tests for buffer size limit
Adding tests for pio_set_buffer_size_limit() to implicitly
flush data.

This buffer size limit decides when data is flushed from
io procs. The buffer size limit is the size of buffer cache
(in bytes) on the io procs (once that limit is reached
data is written out to disk)

* Test that sets buffer size limit to 0
* Test that sets buffer size limit so that data is flushed
  every 1/2/3 vars