Init/Finalize GPTL only once for an app
GPTL initialize and finalize should be only called once for an
application. However PIO init (init_intracomm, ...) and finalize
can be called multiple times, once for each iosystem.

So now we keep track of the number of times user tries to
init/finalize GPTL and only call the GPTL framework once for the
application.