Close GPTL timers on err
Closing some GPTL timers when the functions fail (returns an err).

Without these changes PIO_finalize() fails when GPTL timing summary
is collected in PIO (PIO_ENABLE_INTERNAL_TIMING is enabled). The
GPTLfinalize() function fails and prints out warning messages
related to corrupted internal data structures (due to these timers
that are not closed). These failures occur in the following tests,

* ncdf_fail
* ncdf_simple_tests
* pio_iosystem_tests*

The failures in the first two tests are related to creating a file
that is already created.
The failures in pio_iosystem_tests* is related to an unconditional
sync during close (in PIO) while in define mode (a valid case).