Wait for pending reqs iff we have vars with reqs
Don't wait for pending requests if there are no variables with
pending requests.

'maxreq' is the index of the last vdesc with pending requests
and is used to iterate through the list of vdescs for the file.
Initializing it to -1 to prevent iterating over vdescs when
there are no vdescs with pending requests on the file.