Updating PIO2 C and Fortran interfaces related to inq_varname
PIOc_inq_varname interface does not know the size of the user buffer
passed to it. Without this change, when the buffer has a size less
than PIO_MAX_NAME, strncpy() in PIOc_inq_var can cause overflow on it.

PIOc_inq_varname and PIOc_inq_var are updated to get the buffer size,
and use that information for strncpy().

Fortran interface PIO_inq_varname is updated to consume this change.