Adding test with multiple syncs
Adding a test with multiple sync between writes and reads.

PIO_syncfile() is called after a certain number of calls (the
stride between the syncfile calls increase after every call to
syncfile, currently the stride increases by 1) and then data
is read back (for the variables that have not been tested/read
yet) to verify the value.