Flush if cache size >= pio_buffer_size_limit
We now flush if the current memory allocated for PIO write cache
is >= pio_buffer_size_limit.

In the case we use PIO_USE_MALLOC we now initialize the bget pool
size to pio_cnbuffer_limit.

For cases where we bypass bget memory management (PIO_USE_MALLOC)
and where we use bget memory management (!PIO_USE_MALLOC) we now
have the same logic to decide whether to flush the PIO write
cache. And this logic is now moved to a separate function.

* The current memory allocated is queried using bstats() and if
  it is >= pio_buffer_size_limit the write cache is flushed.
* The old logic of flushing the write cache if it is close to
  being filled/exhausted is retained (this logic is used to
  determine if write cache needs to be flushed if current
  memory allocated is < pio_buffer_size_limit)

Fixes #86