Adding a compile-time setting on maximum cached IO regions
When using PIO with PnetCDF + SUBSET rearranger the number
of non-contiguous regions cached in a single IO process can
grow to a large number. PnetCDF is not efficient at handling
very large number of regions (sub-array requests) in the
data written out. We typically run out of memory or the
write is very slow.

So we need to be able to set a limit on the potential
(after rearrangement) maximum number of non-contiguous
regions cached in an IO process.