Support creating unique iodescs
Supporting creating unique iodescs in a communication domain

The interface of the util function to add an iodesc to the global
list of iodescs now accepts an MPI communicator. The function now
guarantees that iodescs (ids) are unique across the communication
domain specified by the communicator.

This is specifically required for asynchronous I/O as a service
communication since iodesc ids are passed across multiple/disjoint
communication domains.

Also simplifying the logic to find the next iodesc id. We no longer
reuse iodesc ids.