Support creating unique files
Support creating unique files (file desc structures with unique
ids) across a communication domain.

The interface for the util function to add a file to the global
file list now accepts an MPI communicator. The function now
guarantees that files (ids) are unique across the communication
domain specified by the communicator.

This is specifically required for asynchronous I/O as a service
communication since file ids are passed across multiple/disjoint
communication domains.

The util function now returns the file id (similar to util
functions that add iodescs and iosystems to global lists). So
all the logic to find the next pio ncid is now moved inside this
util function.