Get SUBSET rearr working for async i/o service
The SUBSET rearranger splits the union comm of I/O procs and
compute procs. When using asynchronous I/O service, the I/O and
compute comms are disjoint and hence we need to split the union
of the comms, the union_comm.

Now the asynchronous I/O service model works with the SUBSET
rearranger.