Adding PIO_MSG_EXIT and PIO_MSG_INVALID msgs
Adding a PIO_MSG_EXIT message that signals an exit and a
PIO_MSG_INVALID message to initialize a message to an invalid
message.

The loop and the logic is slightly reordered to make the code
more readable. The message id now conveys the intent better and
we no longer modify (and remodify) messages from compute
processes.