Make rearranger optional in the fortran interface
Make rearranger optional in the fortran interface (the default is
the box rearranger). The order of arguments to the PIO_init()
interface was changed so that all optional arguments are at the
end of the list of arguments.

This change ensures backward compatibility with the older
PIO_init() (when using async I/O service) interface.

Updated the tests accordingly.