Adding an option to specify which decomps to save
Adding an option PIO_SAVE_DECOMPS_REGEX that can be modified by
the user to specify the decompositions that need to be saved.

The necessary interfaces are added (including a C++ source file)
but currently do not contain the required logic to parse regex
from user. The new interface is used when user creates a
decomposition to decide whether to save it.