Adding logic to evaluate user reg exp
Adding the necessary logic to evaluate the user-specified
regular expression to save I/O decompositions.

The tokenized string is converted to a postfix expression and
stored. This stored expression is used to decide (by evaluating
the expression) whether to save (based on whether the regular
expression matches the ioid, variable name and file name) the
I/O decomposition.

Fixes #115