Removing an error code used to test PIOc_strerror
In test_pioc.c, check_strerror_netcdf and check_strerror_pio
use a special error code (NC4_LAST_ERROR - 1) to test API
PIOc_strerror.

If PIO is built with PnetCDF, this error code will be passed
to ncmpi_strerror. The expected message is "Unknown Error" but
the returned message might be a known error instead (from newer
versions of PnetCDF).

The issue is that NC4_LAST_ERROR constant may or may not have
the same value in NetCDF and PnetCDF. Due to this uncertainty,
we had better not test PIOc_strerror with "NC4_LAST_ERROR - 1".