Removing cmake logic for C++11 std
Removing cmake logic to enable C++11 . The cmake macro support
is not available for all compiler+platforms as of cmake 3.1 (this
support might be available for most by 3.6, but is not very
usable across compiler+platform combinations unless cmake provides
an easily accessible support matrix).

Directly setting the compiler flags to enable C++11.