Fix check_netcdf usage in PIO
check_netcdf is a collective call that accepts a NetCDF error
code and processes it using the appropriate PIO error handler.
The function returns a PIO error code that needs to be used by
the caller. The previous code, incorrectly, always broadcasted the
error code from the ioroot to all processes (instead of handling
it using the PIO error handler).

* Now using check_netcdf to process a NetCDF error code and using
  the PIO error code returned by the function (rather than
  bcasting error code all the time)
* Fixed some cases where the error code from a PIOc_* function
  was bcasted to all processes (PIOc_* functions already handle
  error codes internally using the appropriate PIO error handlers)
* Initialized uninitialized return code variables. Note that in
  compute processes, for certain iotypes, there are no calls to
  any NetCDF functions (hence the related error codes are not set)