Updating some PIOc_inq* functions to return a soft failure
A failure during a PIO inquire function should not be considered
fatal in most cases.

Certain inquire function failures are fatal, though. For example,
inquiring the name/size of a type (PIOc_inq_type()).

For a soft (non-fatal) failure, the error code is broadcast to
all processes and is directly returned to the user.

We no longer call pio_err() on a soft failure, as it aborts the
program if the error handler is set to PIO_INTERNAL_ERROR.