Optimizing box_rearrange_create on index conversion
Slow integer divisions are used to convert a 1-D index into a
coordinate value. For each element of the data array on the
compute task, the conversion should occur only once.

gcoord_map is used to save all conversion results to avoid
duplicate computation.

Also, we no longer need to process a data element if an IO
task has already been found for it.