Change order of linking pioc with piof
Changing the order in which pioc is added as a dependency for
piof.

The CMake scripts perform detection for NetCDF fortran libraries
when building piof. However NetCDF fortran libraries depend on
NetCDF C libraries and the detection of NetCDF C libraries is
performed when building pioc. So adding pioc as a dependency for
piof after the NetCDF Fortran libraries are detected so that the
NetCDF C libraries are included (since piof depends on pioc and
pioc depends on NetCDF C libraries) in the link order after the
NetCDF Fortran libraries.

Without this change builds with separate NetCDF C and Fortran
library installation fails (at least with intel 18.0.4) since
NetCDF C libraries are included in the link order before the
NetCDF Fortran libraries.

Fixes #136