Remove unused varlist2 struct
Removing the var linked list, varlist2, that is not used by
the current code. This change was added to remove limits on the
number of variables in a file, however it was decided that the
performance impact (of moving from an array to linked list) was
not worth the change. So we do not use varlist2.

Fixes #177