Inquire varids in C unit tests
Several C unit tests assumed varids instead of inquiring them
everytime a file is re-opened (the tests assumed the varids
remained the same after close + reopen of a file).

Some PIO micro timers are created when the user defines/inquires
a variable. If the user does not inquire varid after reopening
a file micro timers for that variable is left in an uninitialized
state and results in an error when PIO_MICRO_TIMING is enabled.