Remove global compute node buffer block
Since we provide BGET with an allocator function a separate
global compute node buffer block (CN_bpool) is not required.
BGET allocates the buffer block the first time bget() is called
(by calling the user-specified allocator) and adds it to the
buffer pool, so a separate global buffer block to initialize
the buffer pool is not required.