Set pio_cnbuffer_limit to pio_buffer_size_limit
Instead of setting pio_cnbuffer_limit to 32MB (that cannot be
changed), we now set it based on the value of
pio_buffer_size_limit. The PIO buffer size limit can be set by
the user and hence should allow users to determine the size of
the compute node bget buffer increment.