Disable modifying user iotype
When SCORPIO is unable to open a file with PnetCDF it internally
retries by switching the iotype to NetCDF.

Unfortunately this causes some E3SM model components to reset the
iotype, if a read of a NetCDF4 file fails with PnetCDF, to NetCDF.
So all subsequent reads and writes use the NetCDF iotype adversely
affecting the performance. So disabling modifying the user type
for now.