Adding custom test targets iff testing is enabled
Adding the custom test targets, "tests" and "check", only if
testing is enabled (PIO_ENABLE_TESTS is turned on).

Also changing the default for PIO_ENABLE_TESTS from ON to OFF
(Now by default tests are not built with Scorpio)

Note: Since "tests" is a generic target name it causes issues
with applications, e.g. HOMME, that include Scorpio within the
application build system (without using the newer cmake features
to add the library as an external project) and have the same
target.