Setting ROMIO hint romio_no_indep_rw based on MPICH version
For lustre file system, there is a confirmed bug with MPICH 3.2.1,
which might cause a division-by-zero SIGFPE inside ROMIO's lustre
driver on some non-aggregator processes with a zero stripe count.

To avoid potential division-by-zero exceptions, we choose not to
set romio_no_indep_rw to "true" for MPICH 3.2.1 or lower versions.