Enabling ROMIO hint romio_no_indep_rw for Cray MPICH and Intel MPI
For MPI implementations based on MPICH 3.2, the division-by-zero
SIGFPE inside ROMIO's lustre driver is only reproducible with
MVAPICH2 so far. It is not reproducible with Cray MPI and Intel
MPI. It is likely that these vendors have applied some fixes.

Therefore, we can still set romio_no_indep_rw to "true" for Cray
MPICH and Intel MPI.