Updating default_subset_partition() to improve the load balancing
When the number of compute tasks is not divisible by the number of
IO tasks, the existing algorithm assigns most or all of the extra
compute tasks to the last IO task. This might cause performance
issues for SUBSET rearranger.

The new algorithm assigns the extra compute tasks evenly to some
IO tasks with lower ranks to achieve a better load balancing.