Reopening test outputs with the write option
In the case where we sync data to disk by close+open (e.g. for
the ADIOS iotype) reopening the output files that needs to be
modified with the write option.

Without this change the ADIOS tests fail because the files that were
being modified multiple times were reopened with the read only
option