Avoid timer recursion in ADIOS close
Avoid recursion in GPTL timers while closing ADIOS files. If
conversion to NetCDF (from ADIOS) is enabled, closing an ADIOS
file results in the conversion library being called and the
conversion routine calls PIOc_closefile again (for closing the
converted NetCDF file using PnetCDF/NetCDF).

To avoid these recursive timers we separate timers for ADIOS and
non-ADIOS closefiles (we already have specific timers for ADIOS
closefile). Although recursion here is not an error, this change
helps in debugging of timers that are not stopped correctly.