Assert in GPTL on recursion
Add an option in GPTL to assert on recursive timers. Since we do
not have recursive timers in Scorpio (only PIOc_closefile is
recursive, but in this function timers are not recursive) this
option allows us to find timers that are not stopped correctly

This option is only enabled when Scorpio internal timing is turned
on and we use the internal GPTL library.