Fixing IBM XL compiler errors on some multiline prints
In PR #337, some long print messages were replaced with prints
that span multiple lines. For IBM XL compiler, there are build
errors on some multiline prints:
1515-022 (S) Syntax Error: Extra token XXXX was found. The token
is ignored.

This is not reproducible with GNU/Intel/PGI compilers. To fix the
build errors for XL compiler, a comma is required at the end of
each double-quoted substring.