Put var test now writes at idx to write a substr
This test case now writes out strings at variable indices of a 1d
string array and then modifies a specified suffix (modifies from
an index of each string) of each string. The user string written
out modifies a substring, determined by the length of the user
specified string, of the current contents of the strings in
the variable.