Saving adios_type to convert put_var variables correctly
This patch is provided by Tahsin Kurc, which performs expected
data type conversion for PIOc_put_vars using ADIOS IO type.

Because buffers of different data types are merged, the adios_type
of the data buffer has to be saved during write in order to do the
conversion correctly later.