Accounting for the variable type conversion in the I/O stats
The I/O stats should be closest to what the application developer
expects. It is possible that a SCORPIO application uses a double
type decomposition and data buffer to write float type variables
to an output NetCDF file. In this case, we should use the type
of that variable to calculate bytes read or written.

For PIOc_write_darray/PIOc_read_darray, we can inquire var info
on pio_type and type_size and cache them in file->varlist.