Fix crash when fortran buffer is large during get
When reading strings from a file if the fortran buffer is larger
than the number of strings in the variable ensure that we stop
copying data. The Fortran buffer provided by the user can be
larger than the number of strings in the variable and without this
fix Scorpio accesses memory outside (array index out of bounds)
the C string array when converting C strings to Fortran strings,
leading to a crash.

We warn the user now in this case (and abort the copying of
strings)