Avoiding abort in Fortran lib
Since the error handlers set by the user is only being handled
in the C library avoid aborting when querying the size of a
variable fails in the Fortran library. A warning is printed out
instead and we let the C library handle the error.

(e.g. Trying to get a variable that is not defined in the file etc)