Fix puts to read only file in ls test
In the newly added test (in PR #417) when files are closed and
reopened for syncing data (this is only done for the ADIOS iotype
currently) they were incorrectly opened as read only files and
variables were being written to it.

Ensure that the file is opened in write mode the second time
before writing data to it.