Init varid in vardesc
The varid in the vardesc structure was not being initialized
inside the library.

This issue did not show up until PR #405 (@979da35) where the
varid was initialized to -1 (Before the change in PR #405 varid
was being set to 0). Also the varid field in the vardesc structure
is only used when micro timers are enabled.

This change initializes the varid correctly when the user defines
a variable or inquires its id.

Without this change tests fail (invalid memory access during sync)
when micro timers are enabled (after PR #405)