Adding support for args to tests
Adding support for command line arguments that can be used by
the Fortran tests.

The command line arguments that are arguments to tests need to
have the "--pio-tf-targ" prefix. All command line arguments with
this prefix are parsed and cached by the testing framework when
parsing the command line arguments. The unit test can access the
cached arguments using the PIO_TF_Get_test_arg() function.

Note that the user is responsible for converting the argument
values from strings to appropriate types.

We currently support up to 10 arguments for tests.