More info when aborting with PIO_INTERNAL_ERROR
Adding the filename in the error message when createfile or
openfile fails with the PIO_INTERNAL_ERROR error handler.

Many components in E3SM use the PIO_INTERNAL_ERROR error handler
and without this change the error message only includes the
error message from the NetCDF library (generic error message for
the error code). Including the filename in the error message
gives more information to the user about the failure.

(Note: We are now passing temporary file handles to the
 check_netcdf function)