Move bcast of file mode before check_netcdf
When the error handler is set to PIO_RETURN_ERROR when the call
to the underlying I/O library fails only the I/O processes get the
error code from the I/O library. This can cause a hang on the
non I/O processes if the file mode is bcasted after the
check_netcdf call.