Reorganize the contents of CMakeLists.txt
Reorganize the CMake code in CMakeLists.txt so that its more
readable & maintainable. Also adding more documentation in the
CMakeLists.txt files.

No new CMake code was added in this commit. The existing CMake
code was moved around and more documentation was added. Also no
CMake code was removed in this commit.