Fix check for MPI_Comm_f2c in gptl
MPI_Comm_f2c is frequently declared as a macro in the MPI header
file. So using cmake to check for the symbol instead of a function
when checking for availability of MPI_Comm_f2c

Also adding the necessary MPI headers/libs needed for the check &
a generic message indicating that the GPTL library is being
configured.