Adding extra padding in NetCDF output file header
NetCDF and PnetCDF by default do not reserve any extra space in
the NetCDF file headers (compactness etc). However this can be a
problem (performance related) when renaming variables etc since
this information is stored in the NetCDF header. Adding more
information into the header requires the libraries to resize the
file, extend the header space and copy the contents of the file
(similar to realloc).

So reserving some extra space in the header when creating NetCDF
files by replacing current calls to nc_enddef()/ncmpi_enddef()
with nc__enddef()/ncmpi__enddef() that allow reserving extra
space in the header.