Check Fortran compiler name instead of C compiler name
There is some logic in the CMakeLists.txt file that checks the name of
the C compiler to determine if it is the GNU compiler. This logic should
be changed to check the name of the Fortran compiler instead. This is
important on a Mac where the C compiler is clang and the Fortran compiler
is gfortran.